+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # cuGraph 0.14.0 (Date TBD)
 
 ## New Features
+- PR #756 Add Force Atlas 2 layout
 - PR #822 Added new functions in python graph class, similar to networkx
 - PR #840 OPG degree
+- PR #875 UVM notebook 
 
 ## Improvements
 - PR #876 Update BFS with rmm::device_vector and add c++ tests
@@ -35,6 +37,9 @@
 - PR #866 Use RAII graph class in KTruss
 - PR #867 Updates to support the latest flake8 version
 - PR #874 Update setup.py to use custom clean command
+- PR #878 Updated build script
+- PR #879 Add docs build script to repository
+- PR #818 Initial version of new "benchmarks" folder
 
 ## Bug Fixes
 - PR #763 Update RAPIDS conda dependencies to v0.14